Modularize and add support for multiple arguments
Move the executing code from `index.js` into a function and export it as
a module. Add `bin/scf.js` as an executable which will require the
module and run the function.
This also adds the ability to call the program with multiple URLs (as an
array, which is standard for process.argv) or, if one wishes, with a
single string, for simpler use as a module.